backend/requirements.txt
# CrewAI and Agents
crewai>=0.51.0
crewai-tools>=0.8.0

# Groq AI
groq>=0.4.0

# FastAPI Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.9

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Vector Store for Memory
chromadb>=0.4.22

# Image Processing
pillow>=10.2.0
numpy>=1.26.0

# Environment
python-dotenv>=1.0.0

# Utilities
structlog>=24.1.0
